Refactor to helper methods
